Separated TensorProduct
